# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dental CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{patientId}/messages:
    get:
      operationId: listMessages
      summary: List messages for a patient
      security:
        - cookieAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
    post:
      operationId: sendMessage
      summary: Send WhatsApp message to patient
      security:
        - cookieAuth: []
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    MessageDirection:
      type: string
      enum: [inbound, outbound]
    Message:
      type: object
      properties:
        id:
          type: string
        clinicId:
          type: string
        patientId:
          type: string
        direction:
          $ref: "#/components/schemas/MessageDirection"
        senderId:
          type: string
          nullable: true
        content:
          type: string
        whatsappMessageId:
          type: string
          nullable: true
        isRedAlert:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clinicId
        - patientId
        - direction
        - senderId
        - content
        - whatsappMessageId
        - isRedAlert
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 4096
      required:
        - content
    MessagesResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
          required:
            - messages
      required:
        - success
        - data
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            message:
              $ref: "#/components/schemas/Message"
          required:
            - message
      required:
        - success
        - data

    # ─── Notifications ────────────────────────────────────────────────────────
...